==> 01_downloadPictureFromYahoo.ipynb <==
# yahoo画像検索でキーワード検索&好きか嫌いかの教師データを与える
---
### 利用バージョン
python 3

## 概要
ここでは
1. SEARCH_LISTに検索キーワードを入れる。
1. 各キーワードが好き(1)か嫌い(0)か回答します。
1. それぞれのキーワードで画像検索し画像ファイルをダウンロードします。また各画像が好き(1)か嫌い(0)かのリストを作成します。　　

といった事をします。

yahoo画像検索で画像ダウンロードする機能は、[ここ](http://karaage.hatenadiary.jp/entry/2017/08/23/073000)のものを利用させてもらっています。


## 変数定義
* yahoo画像検索をスクレイピングする際に使用するメールアドレスをMY_EMAIL_ADDRに記載します。
* SEARCH_LISTに検索するキーワードを入力します。

MY_EMAIL_ADDR = 'mori5602@gmail.com'
SEARCH_LIST = [u'猫', u'犬']

## 教師データ設定
* 以下のセルを実行すると、各キーワードが好きかどうか聞かれるので答えます。
* デモではシンプルに、犬が好き(1)、猫が嫌い(0)で説明します。

CHOICE_LIST = []

for i in SEARCH_LIST:
    loopFlag = True
    while loopFlag == True:
        choice = input(u'do you like {} ?[y/N]'.format(i)).lower()

        if choice in ['y', 'ye', 'yes']:
            CHOICE_LIST.append(1)
            loopFlag = False
        elif choice in ['n', 'no']:
            CHOICE_LIST.append(0)
            loopFlag = False
    
print (CHOICE_LIST)

## 画像ダウンロード＆リスト作成
以下のセルを実行すると、次の２つの結果が得られます。
* dataset/01_downloadPictに画像がダウンロードされます。
* dataset/01_download_pict_anser.csvに、ダウンロードした画像ファイル名と好き(1)or嫌い(0)の一覧が出力されます。

# coding: utf-8

import os
import sys
import traceback
from mimetypes import guess_extension
from time import time, sleep
from urllib.request import urlopen, Request
from urllib.parse import quote
from bs4 import BeautifulSoup
import pandas as pd
import numpy as np

class Fetcher:
    def __init__(self, ua=''):
        self.ua = ua

    def fetch(self, url):
        req = Request(url, headers={'User-Agent': self.ua})
        try:
            with urlopen(req, timeout=3) as p:
                b_content = p.read()
                mime = p.getheader('Content-Type')
        except:
            sys.stderr.write('Error in fetching {}\n'.format(url))
            sys.stderr.write(traceback.format_exc())
            return None, None
        return b_content, mime

fetcher = Fetcher(MY_EMAIL_ADDR)

def remove_dir_and_file(top):
    for root, dirs, files in os.walk(top, topdown=False):
        for name in files:
            os.remove(os.path.join(root, name))
        for name in dirs:
            os.rmdir(os.path.join(root, name))

def fetch_and_save_img(key, word, anser):
    df = pd.DataFrame(columns=['pictFileName', 'anser'])
    
    data_dir = os.path.join(u'dataset', u'01_downloadPict')
    if not os.path.exists(data_dir):
        os.makedirs(data_dir)

    for i, img_url in enumerate(img_url_list(word)):
        sleep(0.1)
        img, mime = fetcher.fetch(img_url)
        if not mime or not img:
            continue
        ext = guess_extension(mime.split(';')[0])
        if ext in ('.jpe', '.jpeg'):
            ext = '.jpg'
        if not ext:
            continue
            
        pictFileName = u'{0}_{1}{2}'.format(key, i, ext)
        result_file = os.path.join(data_dir, pictFileName)

        df_tmp = pd.Series([os.path.join(u'dataset', u'01_downloadPict', pictFileName), anser], index=['pictFileName', 'anser'])
        df = df.append(df_tmp, ignore_index=True)
        
        with open(result_file, mode='wb') as f:
            f.write(img)
        print('fetched', img_url)

    return df, data_dir

def img_url_list(word):
    """
    using yahoo (this script can't use at google)
    """
    url = 'http://image.search.yahoo.co.jp/search?n=60&p={}&search.x=1'.format(quote(word))
    byte_content, _ = fetcher.fetch(url)
    structured_page = BeautifulSoup(byte_content.decode('UTF-8'), 'html.parser')
    img_link_elems = structured_page.find_all('a', attrs={'target': 'imagewin'})
    img_urls = [e.get('href') for e in img_link_elems if e.get('href').startswith('http')]
    img_urls = list(set(img_urls))
    return img_urls

if __name__ == '__main__':
    outfilePath = os.path.join(u'dataset', u'01_downloadPict')
    remove_dir_and_file(outfilePath)
                        
    df = pd.DataFrame(columns=['pictFileName', 'anser'])
    for key, (word, anser) in enumerate(zip(SEARCH_LIST, CHOICE_LIST)):
        df_tmp, data_dir = fetch_and_save_img(key, word, anser)
        df = df.append(df_tmp, ignore_index=True)
    
    # リストをランダムソートしてcsvに出力
    df = df.reindex(np.random.permutation(df.index))
    df.to_csv(os.path.join(u'dataset', '01_download_pict_anser.csv'), index=None, header=None, sep= ' ')
