==> 02_reSizePicture.ipynb <==
# 画像サイズを統一する
---
### 利用バージョン
python 3

## 概要
ここでは、
* リサイズ後のサイズを指定します。
* 指定したサイズに合わせてリサイズします。

といった事をします。

## 変数定義
* IMG_SIZEにリサイズしたい解像度を指定します。

IMG_SIZE = 250

## リサイズ実行
* dataset/02_resizePictにリサイズした画像ファイルを出力します。
* ataset/02_resize_pict_anser.csvに、リサイズした画像ファイル名と好き(1)or嫌い(0)の一覧が出力されます。

# -*- coding:utf-8 -*-

import pandas as pd
import numpy as np
from PIL import Image
from glob import glob
from os.path import join, relpath
import random
import os

#特定フォルダ配下にあるファイルを全削除する関数
def remove_dir_and_file(top):
    for root, dirs, files in os.walk(top, topdown=False):
        for name in files:
            os.remove(os.path.join(root, name))
        for name in dirs:
            os.rmdir(os.path.join(root, name))
            
def main():
    # 入力データ、出力データを格納するパスを指定
    infilePath = join(u'dataset', u'01_downloadPict')
    outfilePath = join(u'dataset', u'02_resizePict')

    # 出力フォルダにある画像ファイルを全部削除します。
    remove_dir_and_file(outfilePath)

    # 画像ファイルと好き(1) or 嫌い(0)のリストをデータフレームに読み込み
    df_pictAnser = pd.read_csv(join(u'dataset', u'01_download_pict_anser.csv'), names=[u'pictFileName', 'anser'], sep=' ')



    # 出力する画像ファイルと好き or 嫌い　のリストを格納するデータフレーム用意
    df_reSizePictAnser = pd.DataFrame(columns=df_pictAnser.columns)

    # 画像ファイル分、以下を繰り返す
    for key, row in df_pictAnser.iterrows():
        infilePathName = row['pictFileName']
        infileName = os.path.basename(infilePathName)
        outfilePathName = join(outfilePath, infileName)

        print (u'resize : {0} to {1}'.format(infileName, outfilePathName)) 
        pflag = False

        # 画像ファイルとして開く&リサイズできるか試す
        try:
            pictFile = Image.open(infilePathName)
            reSizePictFile = pictFile.resize((IMG_SIZE, IMG_SIZE))
            pflag = True
        except:
            print ('cannot open {0}'.format(infilePathName))

        if pflag == True:
            try:
                # RGB形式ではない場合、RGB形式に変換する。
                if reSizePictFile.mode != 'RGB':
                    reSizePictFile = reSizePictFile.convert('RGB')

                # リサイズしたファイルを出力する。
                reSizePictFile.save(outfilePathName)

                se_tmp = pd.Series(data=[outfilePathName, row[u'anser']], index=['pictFileName', 'anser'])
                df_reSizePictAnser = df_reSizePictAnser.append(se_tmp, ignore_index=True)

            except IOError:
                print ('cannot convert {0}'.format(fi))

    # リサイズ後の画像ファイル名と好き(1)or 嫌い(0)　の一覧をランダムソートして出力する。
    df_reSizePictAnser = df_reSizePictAnser.reindex(np.random.permutation(df_reSizePictAnser.index))
    df_reSizePictAnser.to_csv(join(u'dataset', '02_resize_pict_anser.csv'), index=None, header=None, sep= ' ')
    
if __name__ == '__main__':
    main()
