==> 03_duplicatePictureFile.ipynb <==
# 画像を水増しする
---
### 利用バージョン
python 3

## 概要
ここでは、
* 各画像、コピーして複製する回数を変数に設定します。
* 同じファイルをコピーして画像ファイルを水増しします。

といった事をします。

## 変数定義
* 画像ファイルのコピー回数をDUPLICATE_FILEに設定します。

DUPLICATE_FILE = 20

## コピー実行
以下のセルを実行すると、次の２つの結果が得られます。
* dataset/03_duplicatePictに、コピーして水増しした画像ファイルを格納します。
* dataset/03_duplicate_pict_anser.csvに、水増しした画像ファイル名と好き(1)or嫌い(0)の一覧が出力されます。

# -*- coding:utf-8 -*-
import pandas as pd
from PIL import Image
from glob import glob
import random
import os
import shutil
import pandas as pd
import numpy as np

#特定フォルダ配下にあるファイルを全削除する関数
def remove_dir_and_file(top):
    for root, dirs, files in os.walk(top, topdown=False):
        for name in files:
            os.remove(os.path.join(root, name))
        for name in dirs:
            os.rmdir(os.path.join(root, name))

def main():
    # 出力フォルダを格納するパス指定＆対象ディレクトリ配下にあるファイルを全削除
    outfilePath = os.path.join(u'dataset', u'03_duplicatePict')
    remove_dir_and_file(outfilePath)

    # 画像ファイルと好き(1) or 嫌い(0)のリストをデータフレームに読み込み
    df_pictAnser = pd.read_csv(os.path.join(u'dataset', u'02_resize_pict_anser.csv'), names=[u'pictFileName', 'anser'], sep=' ')

    # 出力する画像ファイルと好き or 嫌い　のリストを格納するデータフレーム用意
    df_duplicatePictAnser = pd.DataFrame(columns=df_pictAnser.columns)

    for key, row in df_pictAnser.iterrows():
        infilePathName = row[u'pictFileName']
        fileName = os.path.basename(infilePathName)
        fi, ext = os.path.splitext(fileName)

        # DUPLICATE_FILE数分ファイルを複製する
        for i in range(DUPLICATE_FILE):
            outfilePathName = os.path.join(outfilePath, u'{0}_{1}{2}'.format(fi, i, ext))
            shutil.copyfile(infilePathName, outfilePathName)
            se_tmp = pd.Series(data=[outfilePathName, row[u'anser']], index=df_duplicatePictAnser.columns)
            df_duplicatePictAnser = df_duplicatePictAnser.append(se_tmp, ignore_index=True)

    # リサイズ後の画像ファイル名と好き(1)or 嫌い(0)　の一覧を出力する。
    df_duplicatePictAnser = df_duplicatePictAnser.reindex(np.random.permutation(df_duplicatePictAnser.index))
    df_duplicatePictAnser.to_csv(os.path.join(u'dataset', '03_duplicate_pict_anser.csv'), index=None, header=None, sep= ' ')

    print (u'水増し後の全ファイル数:{}'.format(len(os.listdir(outfilePath))))
    
if __name__ == '__main__':
    main()
